--- api/.gitignore
@@ -1,6 +1,0 @@
-.vscode/
-node_modules
-# Keep environment variables out of version control
-.env
-prisma/client/
-dist/
