--- api/Justfile
@@ -1,16 +1,0 @@
-set dotenv-filename := ".env"
-
-default:
-      ~/.deno/bin/deno run -A src/index.ts
-
-build:
-     cd ../ && npm run build
-     ~/.deno/bin/deno task generate
-
-check:
-     ~/.deno/bin/deno fmt ./src/**/*.ts
-     ~/.deno/bin/deno lint ./src/**/*.ts
-     ~/.deno/bin/deno check ./src/**/*.ts
-
-test:
-     ~/.deno/bin/deno test -A ./test/vault.ts
