--- api/prisma/schema/models/corner.prisma
@@ -1,14 +1,0 @@
-model Corner {
-  id              String  @id @default(dbgenerated("gen_random_uuid()")) @db.Uuid
-  title           String
-  description     String
-  requesting_mail Boolean
-  mail_schema     Json
-  user_id         String  @db.Uuid
-  user            User    @relation(fields: [user_id], references: [id], onDelete: Cascade)
-  podcast_id      String  @db.Uuid
-  podcast         Podcast @relation(fields: [podcast_id], references: [id], onDelete: Cascade)
-  mails           Mail[]
-
-  @@map("corners")
-}
