--- api/prisma/schema/models/episode.prisma
@@ -1,16 +1,0 @@
-model Episode {
-  id           String   @id @default(dbgenerated("gen_random_uuid()")) @db.Uuid
-  title        String
-  audio_url    String?
-  user_id      String?  @db.Uuid
-  podcast_id   String   @db.Uuid
-  srt_url      String?
-  created_at   DateTime @default(now()) @db.Timestamptz(3)
-  sections     Json
-  description  String?
-  duration_sec Int?
-  podcast      Podcast  @relation(fields: [podcast_id], references: [id], onDelete: Cascade)
-  user         User?    @relation(fields: [user_id], references: [id], onDelete: Cascade)
-
-  @@map("episodes")
-}
