--- api/prisma/schema/models/mail.prisma
@@ -1,11 +1,0 @@
-model Mail {
-  id         String   @id @default(dbgenerated("gen_random_uuid()")) @db.Uuid
-  body       Json
-  user_id    String   @db.Uuid
-  user       User     @relation(fields: [user_id], references: [id], onDelete: Cascade)
-  corner_id  String   @db.Uuid
-  corner     Corner   @relation(fields: [corner_id], references: [id], onDelete: Cascade)
-  created_at DateTime @default(now()) @db.Timestamptz(3)
-
-  @@map("mails")
-}
