--- api/prisma/schema/models/podcast.prisma
@@ -1,13 +1,0 @@
-model Podcast {
-  id          String    @id @default(dbgenerated("gen_random_uuid()")) @db.Uuid
-  title       String
-  user_id     String?   @db.Uuid
-  icon        String    @db.VarChar(1)
-  created_at  DateTime  @default(now()) @db.Timestamptz(3)
-  description String?
-  episodes    Episode[]
-  corners     Corner[]
-  user        User?     @relation(fields: [user_id], references: [id], onDelete: Cascade)
-
-  @@map("podcasts")
-}
