--- api/prisma/schema/models/script.prisma
@@ -1,11 +1,0 @@
-model Script {
-  id          String  @id @default(dbgenerated("gen_random_uuid()")) @db.Uuid
-  template    Json
-  user_id     String  @db.Uuid
-  title       String
-  description String?
-  arguments   Json    @default("{}")
-  user        User    @relation(fields: [user_id], references: [id], onDelete: Cascade)
-
-  @@map("scripts")
-}
