--- api/prisma/schema/models/task.prisma
@@ -1,23 +1,0 @@
-model Task {
-  id                   String     @id @default(dbgenerated("gen_random_uuid()")) @db.Uuid
-  status               TaskStatus
-  args                 Json
-  user_id              String?    @db.Uuid
-  execute_after        DateTime   @default(now()) @db.Timestamptz(3)
-  executed_at          DateTime?  @db.Timestamptz(3)
-  executed_finished_at DateTime?  @db.Timestamptz(3)
-  result               Json?
-  cron                 String?
-  user                 User?      @relation(fields: [user_id], references: [id], onDelete: Cascade)
-
-  @@map("tasks")
-}
-
-enum TaskStatus {
-  PENDING
-  RUNNING
-  COMPLETED
-  FAILED
-
-  @@map("task_status")
-}
