--- api/prisma/schema/models/user.prisma
@@ -1,14 +1,0 @@
-model User {
-  id       String    @id @default(dbgenerated("gen_random_uuid()")) @db.Uuid
-  auth_id  String    @unique
-  email    String    @unique
-  name     String?
-  episodes Episode[]
-  podcasts Podcast[]
-  scripts  Script[]
-  tasks    Task[]
-  corners  Corner[]
-  mails    Mail[]
-
-  @@map("users")
-}
