--- api/prisma/schema/schema.prisma
@@ -1,11 +1,0 @@
-generator client {
-  provider        = "prisma-client-js"
-  output          = "../client"
-  previewFeatures = ["deno", "prismaSchemaFolder"]
-}
-
-datasource db {
-  provider  = "postgresql"
-  url       = env("DATABASE_URL")
-  directUrl = env("DIRECT_DATABASE_URL")
-}
