--- src/deno.d.ts
@@ -1,2 +1,0 @@
-// biome-ignore lint:
-declare const Deno: any;
