+++ tsconfig.json
@@ -4,8 +4,7 @@
 		"useDefineForClassFields": true,
 		"lib": ["ESNext", "DOM", "DOM.Iterable"],
 		"paths": {
-			"@/*": ["./src/*"],
-			"prisma_client": ["./api/prisma/client"]
+			"@/*": ["./src/*"]
 		},
 		"module": "ESNext",
 		"skipLibCheck": true,
@@ -25,6 +24,5 @@
 		"noUncheckedIndexedAccess": true
 	},
 	"include": ["src"],
-	"exclude": ["api"],
 	"references": [{ "path": "./tsconfig.node.json" }]
 }
